// src/module/auth/domain/dto/validate-token-dto.ts
import { IsBoolean, IsNotEmpty, IsOptional, IsString } from 'class-validator';

export interface JwtPayload {
  username: string;
  useremail: string;
  iat?: number;
  expiresIn?: number;
  issuer?: string;
  sub: string;
  audience?: string;
}

export class ValidateTokenDTO {
  @IsBoolean()
  @IsNotEmpty()
  valid!: boolean;

  decoded?: JwtPayload;

  @IsString()
  @IsOptional()
  message?: string;
}